+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.9] - 2019-09-09
+### Added
+- Tests for words truncation/ellipsis
+
+### Fixed
+- Logic to apply truncation based on truncateBy prop
+
 ## [0.0.8] - 2019-09-09
 ### Changed
 - Only adding ellipsis if text was truncated